+++ test/unit/test_log.py
@@ -15,14 +15,18 @@
 # You should have received a copy of the GNU General Public License
 # along with Autosubmit.  If not, see <http://www.gnu.org/licenses/>.
 
+import subprocess
+from pathlib import Path
 from typing import TYPE_CHECKING
 
 import pytest
 
 from autosubmit.log.log import AutosubmitCritical, AutosubmitError, Log
+from autosubmit.log.utils import compress_xz, find_uncompressed_files, is_xz_file
 
 if TYPE_CHECKING:
     from pytest_mock import MockFixture
+
 
 """Tests for the log module."""
 
@@ -74,7 +78,7 @@ def test_log_not_format():
     _send_messages(msg)
 
 
-def test_set_file_retrial(mocker: 'MockFixture'):
+def test_set_file_retrial(mocker: "MockFixture"):
     max_retries = 3
 
     # Make os.path.split raise an exception
@@ -86,3 +90,38 @@ def test_set_file_retrial(mocker: 'MockFixture'):
         Log.set_file("imaginary.log", max_retries=max_retries, timeout=1)
 
     assert sleep.call_count == max_retries - 1
+
+
+def test_compress_xz(tmp_path: Path):
+    test_content = "Test content foo bar"
+
+    test_path = tmp_path.joinpath("test-dir")
+    test_path.mkdir()
+
+    input_file = str(test_path.joinpath("test-input.txt"))
+    with open(input_file, "w") as f:
+        f.write(test_content)
+
+    output_file = str(test_path.joinpath("test-compressed.xz"))
+    compress_xz(input_file, output_file, preset=9, extreme=True)
+
+    assert Path(output_file).exists()
+    assert is_xz_file(output_file)
+    assert len(find_uncompressed_files(str(test_path))) == 1
+
+    # Verify same result with xz
+    output_file2 = input_file + ".xz"
+    subprocess.run(["xz", "-9", "-e", "-k", input_file], check=True)
+
+    assert Path(output_file2).exists()
+    assert is_xz_file(output_file2)
+
+    # Verify both files output_file and output_file2 are equal
+    assert subprocess.check_output(
+        ["xz", "-d", "-c", output_file], text=True
+    ) == subprocess.check_output(["xz", "-d", "-c", output_file2], text=True)
+    assert subprocess.run(["cmp", output_file, output_file2], check=True)
+
+    # Cover unexistent path
+    with pytest.raises(FileNotFoundError):
+        find_uncompressed_files(str(tmp_path.joinpath("unexistent_path")))