+++ CHANGELOG.md
@@ -35,6 +35,7 @@ platform does not support wrappers anymore (it was used for testing).
 - Added mypy and ruff to the CI for the files touched by a change granting a 
   higher quality and preservation of the code #2626 #2621
 - Updated base images of micromamba and debian for security update #2610
+- Added a platform option to compress log files of the job's output during log recovery #2555
 
 ### 4.1.15: Bug fixes, enhancements, and new features